Let DafnyLanguageServer depend on DafnyCore instead of DafnyPipeline, so there is no requirement to have javac available
