+++ Source/DafnyLanguageServer/CounterExampleGeneration/DafnyModelVariable.cs
