+++ Source/DafnyLanguageServer/DafnyLanguageServer.csproj
@@ -37,9 +37,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DafnyDriver\DafnyDriver.csproj" />
-    <ProjectReference Include="..\DafnyServer\DafnyServer.csproj" />
-    <ProjectReference Include="..\DafnyPipeline\DafnyPipeline.csproj" />
+    <ProjectReference Include="..\DafnyCore\DafnyCore.csproj" />
   </ItemGroup>
 
   <ItemGroup>
